requirements.txt
pytz
requests
pyjwt
urllib3
websockets
pandas
matplotlib
numpy
tinkoff-investments